+++ src/snowflake/snowpark/dataframe.py
@@ -2,13 +2,15 @@
 #
 # Copyright (c) 2012-2022 Snowflake Computing Inc. All rights reserved.
 #
+from __future__ import annotations
+
 import copy
 import itertools
 import re
 from collections import Counter
 from functools import cached_property
 from logging import getLogger
-from typing import Any, Dict, Iterable, Iterator, List, Optional, Tuple, Union
+from typing import Any, Iterable, Iterator
 
 import snowflake.snowpark
 from snowflake.connector.options import pandas
@@ -73,7 +75,7 @@ from snowflake.snowpark._internal.type_utils import (
 from snowflake.snowpark._internal.utils import (
     TempObjectType,
     column_to_bool,
-    create_statement_query_tag,
+    create_or_update_statement_params_with_query_tag,
     deprecate,
     generate_random_alphanumeric,
     parse_positional_args_to_list,
@@ -117,7 +119,7 @@ def _generate_prefix(prefix: str) -> str:
     return f"{prefix}_{generate_random_alphanumeric(_NUM_PREFIX_DIGITS)}_"
 
 
-def _get_unaliased(col_name: str) -> List[str]:
+def _get_unaliased(col_name: str) -> list[str]:
     unaliased = []
     c = col_name
     while True:
@@ -131,7 +133,7 @@ def _get_unaliased(col_name: str) -> List[str]:
     return unaliased
 
 
-def _alias_if_needed(df: "DataFrame", c: str, prefix: str, common_col_names: List[str]):
+def _alias_if_needed(df: DataFrame, c: str, prefix: str, common_col_names: list[str]):
     col = df.col(c)
     unquoted = c.strip('"')
     if c in common_col_names:
@@ -141,11 +143,11 @@ def _alias_if_needed(df: "DataFrame", c: str, prefix: str, common_col_names: Lis
 
 
 def _disambiguate(
-    lhs: "DataFrame",
-    rhs: "DataFrame",
+    lhs: DataFrame,
+    rhs: DataFrame,
     join_type: JoinType,
-    using_columns: List[str],
-) -> Tuple["DataFrame", "DataFrame"]:
+    using_columns: list[str],
+) -> tuple[DataFrame, DataFrame]:
     # Normalize the using columns.
     normalized_using_columns = {quote_name(c) for c in using_columns}
     #  Check if the LHS and RHS have columns in common. If they don't just return them as-is. If
@@ -405,15 +407,15 @@ class DataFrame:
 
     def __init__(
         self,
-        session: Optional["snowflake.snowpark.Session"] = None,
-        plan: Optional[LogicalPlan] = None,
+        session: snowflake.snowpark.Session | None = None,
+        plan: LogicalPlan | None = None,
         is_cached: bool = False,
     ) -> None:
         self._session = session
         self._plan = session._analyzer.resolve(plan)
         self.is_cached: bool = is_cached  #: Whether it is a cached dataframe
 
-        self._reader: Optional["snowflake.snowpark.DataFrameReader"] = None
+        self._reader: snowflake.snowpark.DataFrameReader | None = None
         self._writer = DataFrameWriter(self)
 
         self._stat = DataFrameStatFunctions(self)
@@ -433,21 +435,29 @@ class DataFrame:
         return self._stat
 
     @df_action_telemetry
-    def collect(self) -> List["Row"]:
+    def collect(self, *, _statement_params: dict[str, Any] | None = None) -> list[Row]:
         """Executes the query representing this DataFrame and returns the result as a
         list of :class:`Row` objects.
-        """
-        return self._internal_collect_with_tag()
 
-    def _internal_collect_with_tag(self) -> List["Row"]:
+        Args:
+            _statement_params: Extra information that should be sent to Snowflake with query.
+
+        Returns:
+            A list of :class:`Row` objects.
+        """
+        return self._internal_collect_with_tag(_statement_params=_statement_params)
+
+    def _internal_collect_with_tag(
+        self, *, _statement_params: dict[str, Any] | None = None
+    ) -> list[Row]:
         # When executing a DataFrame in any method of snowpark (either public or private),
         # we should always call this method instead of collect(), to make sure the
         # query tag is set properly.
         return self._session._conn.execute(
             self._plan,
-            _statement_params={"QUERY_TAG": create_statement_query_tag(3)}
-            if not self._session.query_tag
-            else None,
+            _statement_params=create_or_update_statement_params_with_query_tag(
+                _statement_params, self._session.query_tag, 3
+            ),
         )
 
     def _execute_and_get_query_id(self) -> str:
@@ -460,12 +470,17 @@ class DataFrame:
         )
 
     @df_action_telemetry
-    def to_local_iterator(self) -> Iterator[Row]:
+    def to_local_iterator(
+        self, *, _statement_params: dict[str, Any] | None = None
+    ) -> Iterator[Row]:
         """Executes the query representing this DataFrame and returns an iterator
         of :class:`Row` objects that you can use to retrieve the results.
 
         Unlike :meth:`collect`, this method does not load all data into memory
         at once.
+
+        Args:
+            _statement_params: Extra information that should be sent to Snowflake with query.
 
         Example::
 
@@ -478,21 +493,26 @@ class DataFrame:
         yield from self._session._conn.execute(
             self._plan,
             to_iter=True,
-            _statement_params={"QUERY_TAG": create_statement_query_tag(3)}
-            if not self._session.query_tag
-            else None,
+            _statement_params=create_or_update_statement_params_with_query_tag(
+                _statement_params, self._session.query_tag, 3
+            ),
         )
 
-    def __copy__(self) -> "DataFrame":
+    def __copy__(self) -> DataFrame:
         return DataFrame(self._session, copy.copy(self._plan))
 
     @df_action_telemetry
-    def to_pandas(self, **kwargs) -> "pandas.DataFrame":
+    def to_pandas(
+        self, *, _statement_params: dict[str, Any] | None = None, **kwargs: Any
+    ) -> pandas.DataFrame:
         """
         Executes the query representing this DataFrame and returns the result as a
         `Pandas DataFrame <https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.html>`_.
 
         When the data is too large to fit into memory, you can use :meth:`to_pandas_batches`.
+
+        Args:
+            _statement_params: Extra information that should be sent to Snowflake with query.
 
         Note:
             1. This method is only available if Pandas is installed and available.
@@ -500,9 +520,14 @@ class DataFrame:
             2. If you use :func:`Session.sql` with this method, the input query of
             :func:`Session.sql` can only be a SELECT statement.
         """
-        if not self._session.query_tag:
-            kwargs["_statement_params"] = {"QUERY_TAG": create_statement_query_tag(2)}
-        result = self._session._conn.execute(self._plan, to_pandas=True, **kwargs)
+        result = self._session._conn.execute(
+            self._plan,
+            to_pandas=True,
+            _statement_params=create_or_update_statement_params_with_query_tag(
+                _statement_params, self._session.query_tag, 2
+            ),
+            **kwargs,
+        )
 
         # if the returned result is not a pandas dataframe, raise Exception
         # this might happen when calling this method with non-select commands
@@ -519,7 +544,9 @@ class DataFrame:
         return result
 
     @df_action_telemetry
-    def to_pandas_batches(self, **kwargs) -> Iterator["pandas.DataFrame"]:
+    def to_pandas_batches(
+        self, *, _statement_params: dict[str, Any] | None = None, **kwargs: Any
+    ) -> Iterator[pandas.DataFrame]:
         """
         Executes the query representing this DataFrame and returns an iterator of
         Pandas dataframes (containing a subset of rows) that you can use to
@@ -527,6 +554,9 @@ class DataFrame:
 
         Unlike :meth:`to_pandas`, this method does not load all data into memory
         at once.
+
+        Args:
+            _statement_params: Extra information that should be sent to Snowflake with query.
 
         Example::
 
@@ -544,13 +574,17 @@ class DataFrame:
             2. If you use :func:`Session.sql` with this method, the input query of
             :func:`Session.sql` can only be a SELECT statement.
         """
-        if not self._session.query_tag:
-            kwargs["_statement_params"] = {"QUERY_TAG": create_statement_query_tag(2)}
         yield from self._session._conn.execute(
-            self._plan, to_pandas=True, to_iter=True, **kwargs
+            self._plan,
+            to_pandas=True,
+            to_iter=True,
+            _statement_params=create_or_update_statement_params_with_query_tag(
+                _statement_params, self._session.query_tag, 2
+            ),
+            **kwargs,
         )
 
-    def to_df(self, *names: Union[str, Iterable[str]]) -> "DataFrame":
+    def to_df(self, *names: str | Iterable[str]) -> DataFrame:
         """
         Creates a new DataFrame containing columns with the specified names.
 
@@ -605,7 +639,7 @@ class DataFrame:
         return self.col(name)
 
     @property
-    def columns(self) -> List[str]:
+    def columns(self) -> list[str]:
         """Returns all column names as a list.
 
         The returned column names are consistent with the Snowflake database object `identifier syntax <https://docs.snowflake.com/en/sql-reference/identifiers-syntax.html>`_.
@@ -631,8 +665,8 @@ class DataFrame:
 
     def select(
         self,
-        *cols: Union[ColumnOrName, Iterable[ColumnOrName]],
-    ) -> "DataFrame":
+        *cols: ColumnOrName | Iterable[ColumnOrName],
+    ) -> DataFrame:
         """Returns a new DataFrame with the specified Column expressions as output
         (similar to SELECT in SQL). Only the Columns specified as arguments will be
         present in the resulting DataFrame.
@@ -675,7 +709,7 @@ class DataFrame:
 
         return self._with_plan(Project(names, self._plan))
 
-    def select_expr(self, *exprs: Union[str, Iterable[str]]) -> "DataFrame":
+    def select_expr(self, *exprs: str | Iterable[str]) -> DataFrame:
         """
         Projects a set of SQL expressions and returns a new :class:`DataFrame`.
         This method is equivalent to ``select(sql_expr(...))`` with :func:`select`
@@ -708,8 +742,8 @@ class DataFrame:
 
     def drop(
         self,
-        *cols: Union[ColumnOrName, Iterable[ColumnOrName]],
-    ) -> "DataFrame":
+        *cols: ColumnOrName | Iterable[ColumnOrName],
+    ) -> DataFrame:
         """Returns a new DataFrame that excludes the columns with the specified names
         from the output.
 
@@ -764,7 +798,7 @@ class DataFrame:
         else:
             return self.select(list(keep_col_names))
 
-    def filter(self, expr: ColumnOrSqlExpr) -> "DataFrame":
+    def filter(self, expr: ColumnOrSqlExpr) -> DataFrame:
         """Filters rows based on the specified conditional expression (similar to WHERE
         in SQL).
 
@@ -793,9 +827,9 @@ class DataFrame:
 
     def sort(
         self,
-        *cols: Union[ColumnOrName, Iterable[ColumnOrName]],
-        ascending: Optional[Union[bool, int, List[Union[bool, int]]]] = None,
-    ) -> "DataFrame":
+        *cols: ColumnOrName | Iterable[ColumnOrName],
+        ascending: bool | int | list[bool | int] | None = None,
+    ) -> DataFrame:
         """Sorts a DataFrame by the specified expressions (similar to ORDER BY in SQL).
 
         Examples::
@@ -883,10 +917,14 @@ class DataFrame:
 
     def agg(
         self,
-        exprs: Union[
-            Column, Tuple[str, str], List[Column], List[Tuple[str, str]], Dict[str, str]
-        ],
-    ) -> "DataFrame":
+        exprs: (
+            Column
+            | tuple[str, str]
+            | list[Column]
+            | list[tuple[str, str]]
+            | dict[str, str]
+        ),
+    ) -> DataFrame:
         """Aggregate the data in the DataFrame. Use this method if you don't need to
         group the data (:func:`group_by`).
 
@@ -976,8 +1014,8 @@ class DataFrame:
 
     def rollup(
         self,
-        *cols: Union[ColumnOrName, Iterable[ColumnOrName]],
-    ) -> "snowflake.snowpark.RelationalGroupedDataFrame":
+        *cols: ColumnOrName | Iterable[ColumnOrName],
+    ) -> snowflake.snowpark.RelationalGroupedDataFrame:
         """Performs a SQL
         `GROUP BY ROLLUP <https://docs.snowflake.com/en/sql-reference/constructs/group-by-rollup.html>`_.
         on the DataFrame.
@@ -994,8 +1032,8 @@ class DataFrame:
 
     def group_by(
         self,
-        *cols: Union[ColumnOrName, Iterable[ColumnOrName]],
-    ) -> "snowflake.snowpark.RelationalGroupedDataFrame":
+        *cols: ColumnOrName | Iterable[ColumnOrName],
+    ) -> snowflake.snowpark.RelationalGroupedDataFrame:
         """Groups rows by the columns specified by expressions (similar to GROUP BY in
         SQL).
 
@@ -1037,11 +1075,10 @@ class DataFrame:
 
     def group_by_grouping_sets(
         self,
-        *grouping_sets: Union[
-            "snowflake.snowpark.GroupingSets",
-            Iterable["snowflake.snowpark.GroupingSets"],
-        ],
-    ) -> "snowflake.snowpark.RelationalGroupedDataFrame":
+        *grouping_sets: (
+            snowflake.snowpark.GroupingSets | Iterable[snowflake.snowpark.GroupingSets]
+        ),
+    ) -> snowflake.snowpark.RelationalGroupedDataFrame:
         """Performs a SQL
         `GROUP BY GROUPING SETS <https://docs.snowflake.com/en/sql-reference/constructs/group-by-grouping-sets.html>`_.
         on the DataFrame.
@@ -1079,8 +1116,8 @@ class DataFrame:
 
     def cube(
         self,
-        *cols: Union[ColumnOrName, Iterable[ColumnOrName]],
-    ) -> "snowflake.snowpark.RelationalGroupedDataFrame":
+        *cols: ColumnOrName | Iterable[ColumnOrName],
+    ) -> snowflake.snowpark.RelationalGroupedDataFrame:
         """Performs a SQL
         `GROUP BY CUBE <https://docs.snowflake.com/en/sql-reference/constructs/group-by-cube.html>`_.
         on the DataFrame.
@@ -1095,7 +1132,7 @@ class DataFrame:
             snowflake.snowpark.relational_grouped_dataframe._CubeType(),
         )
 
-    def distinct(self) -> "DataFrame":
+    def distinct(self) -> DataFrame:
         """Returns a new DataFrame that contains only the rows with distinct values
         from the current DataFrame.
 
@@ -1105,7 +1142,7 @@ class DataFrame:
             [self.col(quote_name(f.name)) for f in self.schema.fields]
         ).agg([])
 
-    def drop_duplicates(self, *subset: Union[str, Iterable[str]]) -> "DataFrame":
+    def drop_duplicates(self, *subset: str | Iterable[str]) -> DataFrame:
         """Creates a new DataFrame by removing duplicated rows on given subset of columns.
 
         If no subset of columns is specified, this function is the same as the :meth:`distinct` function.
@@ -1142,7 +1179,7 @@ class DataFrame:
         self,
         pivot_col: ColumnOrName,
         values: Iterable[LiteralType],
-    ) -> "snowflake.snowpark.RelationalGroupedDataFrame":
+    ) -> snowflake.snowpark.RelationalGroupedDataFrame:
         """Rotates this DataFrame by turning the unique values from one column in the input
         expression into multiple columns and aggregating results where required on any
         remaining column values.
@@ -1187,8 +1224,8 @@ class DataFrame:
         )
 
     def unpivot(
-        self, value_column: str, name_column: str, column_list: List[ColumnOrName]
-    ) -> "DataFrame":
+        self, value_column: str, name_column: str, column_list: list[ColumnOrName]
+    ) -> DataFrame:
         """Rotates a table by transforming columns into rows.
         UNPIVOT is a relational operator that accepts two columns (from a table or subquery), along with a list of columns, and generates a row for each column specified in the list. In a query, it is specified in the FROM clause after the table name or subquery.
         Note that UNPIVOT is not exactly the reverse of PIVOT as it cannot undo aggregations made by PIVOT.
@@ -1221,7 +1258,7 @@ class DataFrame:
             Unpivot(value_column, name_column, column_exprs, self._plan)
         )
 
-    def limit(self, n: int) -> "DataFrame":
+    def limit(self, n: int) -> DataFrame:
         """Returns a new DataFrame that contains at most ``n`` rows from the current
         DataFrame (similar to LIMIT in SQL).
 
@@ -1232,7 +1269,7 @@ class DataFrame:
         """
         return self._with_plan(Limit(Literal(n), self._plan))
 
-    def union(self, other: "DataFrame") -> "DataFrame":
+    def union(self, other: DataFrame) -> DataFrame:
         """Returns a new DataFrame that contains all the rows in the current DataFrame
         and another DataFrame (``other``), excluding any duplicate rows. Both input
         DataFrames must contain the same number of columns.
@@ -1255,7 +1292,7 @@ class DataFrame:
         """
         return self._with_plan(UnionPlan(self._plan, other._plan, is_all=False))
 
-    def union_all(self, other: "DataFrame") -> "DataFrame":
+    def union_all(self, other: DataFrame) -> DataFrame:
         """Returns a new DataFrame that contains all the rows in the current DataFrame
         and another DataFrame (``other``), including any duplicate rows. Both input
         DataFrames must contain the same number of columns.
@@ -1281,7 +1318,7 @@ class DataFrame:
         return self._with_plan(UnionPlan(self._plan, other._plan, is_all=True))
 
     @df_usage_telemetry
-    def union_by_name(self, other: "DataFrame") -> "DataFrame":
+    def union_by_name(self, other: DataFrame) -> DataFrame:
         """Returns a new DataFrame that contains all the rows in the current DataFrame
         and another DataFrame (``other``), excluding any duplicate rows.
 
@@ -1307,7 +1344,7 @@ class DataFrame:
         return self._union_by_name_internal(other, is_all=False)
 
     @df_usage_telemetry
-    def union_all_by_name(self, other: "DataFrame") -> "DataFrame":
+    def union_all_by_name(self, other: DataFrame) -> DataFrame:
         """Returns a new DataFrame that contains all the rows in the current DataFrame
         and another DataFrame (``other``), including any duplicate rows.
 
@@ -1334,8 +1371,8 @@ class DataFrame:
         return self._union_by_name_internal(other, is_all=True)
 
     def _union_by_name_internal(
-        self, other: "DataFrame", is_all: bool = False
-    ) -> "DataFrame":
+        self, other: DataFrame, is_all: bool = False
+    ) -> DataFrame:
         left_output_attrs = self._output
         right_output_attrs = other._output
         right_output_attr_by_name = {rattr.name: rattr for rattr in right_output_attrs}
@@ -1365,7 +1402,7 @@ class DataFrame:
 
         return self._with_plan(UnionPlan(self._plan, right_child._plan, is_all))
 
-    def intersect(self, other: "DataFrame") -> "DataFrame":
+    def intersect(self, other: DataFrame) -> DataFrame:
         """Returns a new DataFrame that contains the intersection of rows from the
         current DataFrame and another DataFrame (``other``). Duplicate rows are
         eliminated.
@@ -1388,7 +1425,7 @@ class DataFrame:
         """
         return self._with_plan(Intersect(self._plan, other._plan))
 
-    def except_(self, other: "DataFrame") -> "DataFrame":
+    def except_(self, other: DataFrame) -> DataFrame:
         """Returns a new DataFrame that contains all the rows from the current DataFrame
         except for the rows that also appear in the ``other`` DataFrame. Duplicate rows are eliminated.
 
@@ -1411,9 +1448,7 @@ class DataFrame:
         """
         return self._with_plan(Except(self._plan, other._plan))
 
-    def natural_join(
-        self, right: "DataFrame", join_type: Optional[str] = None
-    ) -> "DataFrame":
+    def natural_join(self, right: DataFrame, join_type: str | None = None) -> DataFrame:
         """Performs a natural join of the specified type (``joinType``) with the
         current DataFrame and another DataFrame (``right``).
 
@@ -1457,10 +1492,10 @@ class DataFrame:
 
     def join(
         self,
-        right: "DataFrame",
-        using_columns: Optional[Union[ColumnOrName, List[ColumnOrName]]] = None,
-        join_type: Optional[str] = None,
-    ) -> "DataFrame":
+        right: DataFrame,
+        using_columns: ColumnOrName | list[ColumnOrName] | None = None,
+        join_type: str | None = None,
+    ) -> DataFrame:
         """Performs a join of the specified type (``join_type``) with the current
         DataFrame and another DataFrame (``right``) on a list of columns
         (``using_columns``).
@@ -1589,10 +1624,10 @@ class DataFrame:
 
     def join_table_function(
         self,
-        func: Union[str, List[str], TableFunctionCall],
+        func: str | list[str] | TableFunctionCall,
         *func_arguments: ColumnOrName,
         **func_named_arguments: ColumnOrName,
-    ) -> "DataFrame":
+    ) -> DataFrame:
         """Lateral joins the current DataFrame with the output of the specified table function.
 
         References: `Snowflake SQL functions <https://docs.snowflake.com/en/sql-reference/functions-table.html>`_.
@@ -1672,7 +1707,7 @@ class DataFrame:
         )
         return DataFrame(self._session, TableFunctionJoin(self._plan, func_expr))
 
-    def cross_join(self, right: "DataFrame") -> "DataFrame":
+    def cross_join(self, right: DataFrame) -> DataFrame:
         """Performs a cross join, which returns the Cartesian product of the current
         :class:`DataFrame` and another :class:`DataFrame` (``right``).
 
@@ -1703,10 +1738,10 @@ class DataFrame:
 
     def _join_dataframes(
         self,
-        right: "DataFrame",
-        using_columns: Union[Column, List[str]],
+        right: DataFrame,
+        using_columns: Column | list[str],
         join_type: JoinType,
-    ) -> "DataFrame":
+    ) -> DataFrame:
         if isinstance(using_columns, Column):
             return self._join_dataframes_internal(
                 right, join_type, join_exprs=using_columns
@@ -1731,8 +1766,8 @@ class DataFrame:
             )
 
     def _join_dataframes_internal(
-        self, right: "DataFrame", join_type: JoinType, join_exprs: Optional[Column]
-    ) -> "DataFrame":
+        self, right: DataFrame, join_type: JoinType, join_exprs: Column | None
+    ) -> DataFrame:
         (lhs, rhs) = _disambiguate(self, right, join_type, [])
         expression = join_exprs._expression if join_exprs is not None else None
         return self._with_plan(
@@ -1744,7 +1779,7 @@ class DataFrame:
             )
         )
 
-    def with_column(self, col_name: str, col: Column) -> "DataFrame":
+    def with_column(self, col_name: str, col: Column) -> DataFrame:
         """
         Returns a DataFrame with an additional column with the specified name
         ``col_name``. The column is computed by using the specified expression ``col``.
@@ -1770,7 +1805,7 @@ class DataFrame:
         """
         return self.with_columns([col_name], [col])
 
-    def with_columns(self, col_names: List[str], values: List[Column]) -> "DataFrame":
+    def with_columns(self, col_names: list[str], values: list[Column]) -> DataFrame:
         """Returns a DataFrame with additional columns with the specified names
         ``col_names``. The columns are computed by using the specified expressions
         ``values``.
@@ -1822,11 +1857,17 @@ class DataFrame:
         return self.select([*old_cols, *new_cols])
 
     @df_action_telemetry
-    def count(self) -> int:
+    def count(self, *, _statement_params: dict[str, Any] | None = None) -> int:
         """Executes the query representing this DataFrame and returns the number of
         rows in the result (similar to the COUNT function in SQL).
+
+        Args:
+            _statement_params: Extra information that should be sent to Snowflake with query.
+
         """
-        return self.agg(("*", "count"))._internal_collect_with_tag()[0][0]
+        return self.agg(("*", "count"))._internal_collect_with_tag(
+            _statement_params=_statement_params
+        )[0][0]
 
     @property
     def write(self) -> DataFrameWriter:
@@ -1854,16 +1895,17 @@ class DataFrame:
     @df_action_telemetry
     def copy_into_table(
         self,
-        table_name: Union[str, Iterable[str]],
+        table_name: str | Iterable[str],
         *,
-        files: Optional[Iterable[str]] = None,
-        pattern: Optional[str] = None,
-        validation_mode: Optional[str] = None,
-        target_columns: Optional[Iterable[str]] = None,
-        transformations: Optional[Iterable[ColumnOrName]] = None,
-        format_type_options: Optional[Dict[str, Any]] = None,
+        files: Iterable[str] | None = None,
+        pattern: str | None = None,
+        validation_mode: str | None = None,
+        target_columns: Iterable[str] | None = None,
+        transformations: Iterable[ColumnOrName] | None = None,
+        format_type_options: dict[str, Any] | None = None,
+        _statement_params: dict[str, Any] | None = None,
         **copy_options: Any,
-    ) -> List[Row]:
+    ) -> list[Row]:
         """Executes a `COPY INTO <table> <https://docs.snowflake.com/en/sql-reference/sql/copy-into-table.html>`__ command to load data from files in a stage location into a specified table.
 
         It returns the load result described in `OUTPUT section of the COPY INTO <table> command <https://docs.snowflake.com/en/sql-reference/sql/copy-into-table.html#output>`__.
@@ -1913,6 +1955,7 @@ class DataFrame:
             target_columns: Name of the columns in the table where the data should be saved.
             transformations: A list of column transformations.
             format_type_options: A dict that contains the ``formatTypeOptions`` of the ``COPY INTO <table>`` command.
+            _statement_params: Extra information that should be sent to Snowflake with query.
             copy_options: The kwargs that is used to specify the ``copyOptions`` of the ``COPY INTO <table>`` command.
         """
         if not self._reader or not self._reader._file_path:
@@ -1991,10 +2034,16 @@ class DataFrame:
                 cur_options=self._reader._cur_options,
                 create_table_from_infer_schema=create_table_from_infer_schema,
             ),
-        )._internal_collect_with_tag()
+        )._internal_collect_with_tag(_statement_params=_statement_params)
 
     @df_action_telemetry
-    def show(self, n: int = 10, max_width: int = 50) -> None:
+    def show(
+        self,
+        n: int = 10,
+        max_width: int = 50,
+        *,
+        _statement_params: dict[str, Any] | None = None,
+    ) -> None:
         """Evaluates this DataFrame and prints out the first ``n`` rows with the
         specified maximum number of characters per column.
 
@@ -2003,14 +2052,15 @@ class DataFrame:
             max_width: The maximum number of characters to print out for each column.
                 If the number of characters exceeds the maximum, the method prints out
                 an ellipsis (...) at the end of the column.
+            _statement_params: Extra information that should be sent to Snowflake with query.
         """
         print(
             self._show_string(
                 n,
                 max_width,
-                _statement_params={"QUERY_TAG": create_statement_query_tag(2)}
-                if not self._session.query_tag
-                else None,
+                _statement_params=create_or_update_statement_params_with_query_tag(
+                    _statement_params, self._session.query_tag, 2
+                ),
             )
         )
 
@@ -2022,11 +2072,11 @@ class DataFrame:
     def flatten(
         self,
         input: ColumnOrName,
-        path: Optional[str] = None,
+        path: str | None = None,
         outer: bool = False,
         recursive: bool = False,
         mode: str = "BOTH",
-    ) -> "DataFrame":
+    ) -> DataFrame:
         """Flattens (explodes) compound values into multiple rows.
 
         It creates a new ``DataFrame`` from this ``DataFrame``, carries the existing columns to the new ``DataFrame``,
@@ -2091,7 +2141,7 @@ class DataFrame:
             FlattenFunction(input._expression, path, outer, recursive, mode)
         )
 
-    def _lateral(self, table_function: TableFunctionExpression) -> "DataFrame":
+    def _lateral(self, table_function: TableFunctionExpression) -> DataFrame:
         result_columns = [
             attr.name
             for attr in self._session._analyzer.resolve(
@@ -2163,7 +2213,7 @@ class DataFrame:
         total_width = sum(col_width) + col_count + 1
         line = "-" * total_width + "\n"
 
-        def row_to_string(row: List[str]) -> str:
+        def row_to_string(row: list[str]) -> str:
             tokens = []
             if row:
                 for segment, size in zip(row, col_width):
@@ -2187,7 +2237,12 @@ class DataFrame:
         )
 
     @df_action_telemetry
-    def create_or_replace_view(self, name: Union[str, Iterable[str]]) -> List[Row]:
+    def create_or_replace_view(
+        self,
+        name: str | Iterable[str],
+        *,
+        _statement_params: dict[str, Any] | None = None,
+    ) -> list[Row]:
         """Creates a view that captures the computation expressed by this DataFrame.
 
         For ``name``, you can include the database and schema name (i.e. specify a
@@ -2199,6 +2254,10 @@ class DataFrame:
         Args:
             name: The name of the view to create or replace. Can be a list of strings
                 that specifies the database name, schema name, and view name.
+            _statement_params: Extra information that should be sent to Snowflake with query.
+
+        Returns:
+            A list of :class:`Row` objects.
         """
         if isinstance(name, str):
             formatted_name = name
@@ -2212,13 +2271,18 @@ class DataFrame:
         return self._do_create_or_replace_view(
             formatted_name,
             PersistedView(),
-            _statement_params={"QUERY_TAG": create_statement_query_tag(2)}
-            if not self._session.query_tag
-            else None,
+            _statement_params=create_or_update_statement_params_with_query_tag(
+                _statement_params, self._session.query_tag, 2
+            ),
         )
 
     @df_action_telemetry
-    def create_or_replace_temp_view(self, name: Union[str, Iterable[str]]) -> List[Row]:
+    def create_or_replace_temp_view(
+        self,
+        name: str | Iterable[str],
+        *,
+        _statement_params: dict[str, Any] | None = None,
+    ) -> list[Row]:
         """Creates a temporary view that returns the same results as this DataFrame.
 
         You can use the view in subsequent SQL queries and statements during the
@@ -2234,6 +2298,7 @@ class DataFrame:
         Args:
             name: The name of the view to create or replace. Can be a list of strings
                 that specifies the database name, schema name, and view name.
+            _statement_params: Extra information that should be sent to Snowflake with query.
         """
         if isinstance(name, str):
             formatted_name = name
@@ -2247,9 +2312,9 @@ class DataFrame:
         return self._do_create_or_replace_view(
             formatted_name,
             LocalTempView(),
-            _statement_params={"QUERY_TAG": create_statement_query_tag(2)}
-            if not self._session.query_tag
-            else None,
+            _statement_params=create_or_update_statement_params_with_query_tag(
+                _statement_params, self._session.query_tag, 2
+            ),
         )
 
     def _do_create_or_replace_view(self, view_name: str, view_type: ViewType, **kwargs):
@@ -2265,9 +2330,15 @@ class DataFrame:
         )
 
     @df_action_telemetry
-    def first(self, n: Optional[int] = None) -> Union[Optional[Row], List[Row]]:
+    def first(
+        self, n: int | None = None, *, _statement_params: dict[str, Any] | None = None
+    ) -> Row | None | list[Row]:
         """Executes the query representing this DataFrame and returns the first ``n``
         rows of the results.
+
+        Args:
+            n: The number of rows to return.
+            _statement_params: Extra information that should be sent to Snowflake with query.
 
         Returns:
              A list of the first ``n`` :class:`Row` objects if ``n`` is not ``None``. If ``n`` is negative or
@@ -2276,20 +2347,22 @@ class DataFrame:
              results, or ``None`` if it does not exist.
         """
         if n is None:
-            result = self.limit(1)._internal_collect_with_tag()
+            result = self.limit(1)._internal_collect_with_tag(
+                _statement_params=_statement_params
+            )
             return result[0] if result else None
         elif not isinstance(n, int):
             raise ValueError(f"Invalid type of argument passed to first(): {type(n)}")
         elif n < 0:
-            return self._internal_collect_with_tag()
+            return self._internal_collect_with_tag(_statement_params=_statement_params)
         else:
-            return self.limit(n)._internal_collect_with_tag()
+            return self.limit(n)._internal_collect_with_tag(
+                _statement_params=_statement_params
+            )
 
     take = first
 
-    def sample(
-        self, frac: Optional[float] = None, n: Optional[int] = None
-    ) -> "DataFrame":
+    def sample(self, frac: float | None = None, n: int | None = None) -> DataFrame:
         """Samples rows based on either the number of rows to be returned or a
         percentage of rows to be returned.
 
@@ -2305,7 +2378,7 @@ class DataFrame:
         )
 
     @staticmethod
-    def _validate_sample_input(frac: Optional[float] = None, n: Optional[int] = None):
+    def _validate_sample_input(frac: float | None = None, n: int | None = None):
         if frac is None and n is None:
             raise ValueError(
                 "'frac' and 'n' cannot both be None. "
@@ -2327,7 +2400,7 @@ class DataFrame:
         """
         return self._na
 
-    def describe(self, *cols: Union[str, List[str]]) -> "DataFrame":
+    def describe(self, *cols: str | list[str]) -> DataFrame:
         """
         Computes basic statistics for numeric columns, which includes
         ``count``, ``mean``, ``stddev``, ``min``, and ``max``. If no columns
@@ -2401,7 +2474,7 @@ class DataFrame:
 
         return res_df
 
-    def with_column_renamed(self, existing: ColumnOrName, new: str) -> "DataFrame":
+    def with_column_renamed(self, existing: ColumnOrName, new: str) -> DataFrame:
         """Returns a DataFrame with the specified column ``existing`` renamed as ``new``.
 
         Example::
@@ -2455,7 +2528,9 @@ class DataFrame:
         return self.select(new_columns)
 
     @df_action_telemetry
-    def cache_result(self) -> "DataFrame":
+    def cache_result(
+        self, *, _statement_params: dict[str, Any] | None = None
+    ) -> DataFrame:
         """Caches the content of this DataFrame to create a new cached DataFrame.
 
         All subsequent operations on the returned cached DataFrame are performed on the cached data
@@ -2493,6 +2568,9 @@ class DataFrame:
             >>> df3.collect()
             [Row(NUM=1), Row(NUM=2), Row(NUM=3)]
 
+        Args:
+            _statement_params: Extra information that should be sent to Snowflake with query.
+
         Returns:
              A :class:`DataFrame` object that holds the cached result in a temporary table.
              All operations on this new DataFrame have no effect on the original.
@@ -2503,17 +2581,21 @@ class DataFrame:
         )
         self._session._conn.execute(
             create_temp_table,
-            _statement_params={"QUERY_TAG": create_statement_query_tag(2)}
-            if not self._session.query_tag
-            else None,
+            _statement_params=create_or_update_statement_params_with_query_tag(
+                _statement_params, self._session.query_tag, 2
+            ),
         )
         new_plan = self._session.table(temp_table_name)._plan
         return DataFrame(session=self._session, plan=new_plan, is_cached=True)
 
     @df_action_telemetry
     def random_split(
-        self, weights: List[float], seed: Optional[int] = None
-    ) -> List["DataFrame"]:
+        self,
+        weights: list[float],
+        seed: int | None = None,
+        *,
+        _statement_params: dict[str, Any] | None = None,
+    ) -> list[DataFrame]:
         """
         Randomly splits the current DataFrame into separate DataFrames,
         using the specified weights.
@@ -2525,6 +2607,7 @@ class DataFrame:
                 weight is specified, the returned DataFrame list only includes
                 the current DataFrame.
             seed: The seed for sampling.
+            _statement_params: Extra information that should be sent to Snowflake with query.
 
         Example::
 
@@ -2555,7 +2638,7 @@ class DataFrame:
             temp_column_name = random_name_for_temp_object(TempObjectType.COLUMN)
             cached_df = self.with_column(
                 temp_column_name, abs_(random(seed)) % _ONE_MILLION
-            ).cache_result()
+            ).cache_result(_statement_params=_statement_params)
             sum_weights = sum(weights)
             normalized_cum_weights = [0] + [
                 int(w * _ONE_MILLION)
@@ -2574,7 +2657,7 @@ class DataFrame:
             return res_dfs
 
     @property
-    def queries(self) -> Dict[str, List[str]]:
+    def queries(self) -> dict[str, list[str]]:
         """
         Returns a ``dict`` that contains a list of queries that will be executed to
         evaluate this DataFrame with the key `queries`, and a list of post-execution
@@ -2612,7 +2695,7 @@ Query List:
 
         return f"{msg}\n--------------------------------------------"
 
-    def _resolve(self, col_name: str) -> Union[Expression, NamedExpression]:
+    def _resolve(self, col_name: str) -> Expression | NamedExpression:
         normalized_col_name = quote_name(col_name)
         cols = list(filter(lambda attr: attr.name == normalized_col_name, self._output))
         if len(cols) == 1:
@@ -2623,7 +2706,7 @@ Query List:
             )
 
     @cached_property
-    def _output(self) -> List[Attribute]:
+    def _output(self) -> list[Attribute]:
         return self._plan.output
 
     @cached_property
@@ -2639,8 +2722,8 @@ Query List:
     def _convert_cols_to_exprs(
         self,
         calling_method: str,
-        *cols: Union[ColumnOrName, Iterable[ColumnOrName]],
-    ) -> List[Expression]:
+        *cols: ColumnOrName | Iterable[ColumnOrName],
+    ) -> list[Expression]:
         """Convert a string or a Column, or a list of string and Column objects to expression(s)."""
 
         def convert(col: ColumnOrName) -> Expression: